+++ CHANGELOG.md
@@ -4,7 +4,10 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 [Semantic Versioning](http://semver.org/).
 
 ### [Unreleased]
-### Changed
+
+## Changed
+- TransactionHandler will now not roll back if transaction is `STATUS_MARKED_ROLLBACK` or `STATUS_ROLLING_BACK`
+- Save of ProcessedEvent will now throw ProcessedEventTrackingException if eventNumber, source or component are not unique
 - ReplaySingleEvent JMX commands can now take an optional commandRuntimeString of the component name
   to work with MI contexts
 - Configure timeouts for releasing transactional advisory locks in event linking